backend/requirements.txt
pymongo
python-dotenv
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
librosa
tensorflow